panyubi/requirements.txt
asgiref==3.5.2
black==22.8.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
mypy-extensions==0.4.3
numpy==1.21.6
opencv-contrib-python==4.6.0.66
pathspec==0.10.1
pdfrw==0.4
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.2.1
reportlab==3.6.11
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.3.0
urllib3==1.26.12
zipp==3.8.1
